+++ iscc_sct/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.0"
+__version__ = "0.2.2"
 from iscc_sct.options import *
 from iscc_sct.utils import *
 from iscc_sct.code_semantic_text import *